Fixed db and added update to beaches

+++ db/migrate/20200130210214_create_beaches.rb
@@ -2,10 +2,10 @@ class CreateBeaches < ActiveRecord::Migration[6.0]
   def change
     create_table :beaches do |t|
       t.string :name
-      t.integer :temp
-      t.integer :weather
-      t.integer :pollution
-      t.integer :surf
+      t.float :temp
+      t.float :weather
+      t.float :pollution
+      t.float :surf
       t.string :address
 
       t.timestamps
